MonopolyRolling
Monopoly, players roll two six-sided dice to determine their move. If the same value is on
both dice, this is called “rolling doubles,” and it means they go again. Write a function that rolls two six-sided dice and, if the values on both die were
the same,it continues rolling until two non-equal values appear on the two dice. In other
words, the function should behave as follows:
1. Generate two random numbers in the 1 to 6 range.
2. If they are not the same, return the sum of all numbers rolled so far.
3. If they are the same, keep track of the total rolled so far and return to step 1