GuessMovieGame
Write a program for a game where the player gets to guess the movie name given the number of letters in it (pretty much like hangman but with movies)!
The rules are simple, the computer randomly picks a movie title, and shows you how many letters it's made up of. Your goal is to try to figure out the movie by guessing one letter at a time. 
If a letter is indeed in the title the computer will reveal its correct position in the word, if not, you lose a point. If you lose 10 points, game over!